Massive Speed gains with restructuring
This commit is a major restructuring of `get_mpas_patches`. Initial tests point
to at least 48x speed increases for some mesh patch collections.

The majority of the speed up was done by not taking subsets of NetCDF4
variables that are on disk. Using the `[:]` syntax with a
`netcdf4.variables['var'][:]` we can pull the variables off disk, and then
subset them in ram, rather then making discontiguous I/O calls to disk.

We are also able to gain some major speed up by decreasing the operations
within the or loop over cells. We can do this by creating a variable
llVertexOnCell that has shape: (nCells, maxEdges, 2), (where the last dimension
(2) represents the latitude and longitude). Thus, we are able to take advantage
of numpy's array operations and vectorizations capabilities; however, this
probably increases the total memory usage, but this area has not been profiled.